KTOR-4954 CallFailed hook consumes exception in writing body
